% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsnap.R
\docType{package}
\name{ggsnap-package}
\alias{ggsnap-package}
\title{ggsnap: Save a 'ggplot2' Plot using '+' Operator}
\description{
Provides ggsnap(), which saves a 'ggplot2' plot to disk in a '+' chain, acting as a thin, chainable wrapper around ggplot2::ggsave(). Can be called multiple times in a chain to save different snapshots.
}
\author{
\strong{Maintainer}: Or Gadish \email{orgadish@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Or Gadish \email{orgadish@gmail.com} [copyright holder]
}

}
\keyword{internal}
